Initial draft of A2A extension specification.
There are more sections I would like to see fleshed out. Specifically
around mapping AP2 flows onto A2A request/responses. That's where the
real meat is, but I'm missing details and time to write it.

There are a few TODOs in this spec as well. I'd like to be able to link
to well-defined data models for each of the 3 types (IntentMandate,
CartMandate, PaymentMandate).